+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# 0.2.0 #
+
+* added Sputnik::Config
+COMMON LOGGING PER SPICEWEASEL
+COMMON DEBUGGING PER SPICEWEASEL
+
 # 0.1.0 #
 
 * added sputnik-example reference implementation